Add cart management features and refactor DTOs

+++ VTT_SHOP_DATABASE/Repositories/CartRepository.cs
@@ -10,17 +10,17 @@ namespace VTT_SHOP_DATABASE.Repositories
         {
         }
 
-        public async Task<Cart?> GetCartByUserIdAsync(int userId)
+        public async Task<Cart?> GetCartByUserIdAsync(long userId)
         {
             return await base.GetAll().FirstOrDefaultAsync(c => c.UserId == userId);
         }
-        public async Task<Cart?> GetCartWithItemsByUserIdAsync(int userId)
+        public async Task<Cart?> GetCartWithItemsByUserIdAsync(long userId)
         {
             return await base.GetAll()
                 .Include(c => c.Items)
                 .FirstOrDefaultAsync(c => c.UserId == userId);
         }
-        public async Task<bool> CartExistsAsync(int userId)
+        public async Task<bool> CartExistsAsync(long userId)
         {
             return await base.GetAll().AnyAsync(c => c.UserId == userId);
         }
